Pinbot/requirements.txt
alabaster==0.7.6
amqp==1.4.6
anyjson==0.3.3
Babel==2.0
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.18
bleach==1.4
celery==3.1.16
certifi==14.5.14
coverage==3.7.1
Django==1.6.5
django-admin-bootstrapped==2.3.1
django-appconf==0.6
django-compressor==1.4
django-cors-headers==1.1.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.2
django-endless-pagination==2.0
django-mobile==0.4.0
django-model-utils==2.2
django-nose==1.4.1
-e git+https://github.com/runforever/django-notifications.git@a95d983d152ff4e4a3d474aa864b7ec51d31d034#egg=django_notifications_hq-master
django-redis==3.8.0
django-simple-captcha==0.4.3
django-summernote==0.6.8
django-tokenapi==0.2.3
django-websocket-redis==0.4.2
django-xadmin==0.4.2
docutils==0.12
flake8==2.2.2
flower==0.7.3
futures==2.1.6
html5lib==0.999
ipdb==0.8
ipython==2.2.0
jieba==0.35
Jinja2==2.8
jsonfield==1.0.3
kombu==3.0.23
MarkupSafe==0.23
mccabe==0.2.1
mock==1.0.1
mongoengine==0.8.7
MySQL-python==1.2.5
nose==1.3.7
pep8==1.5.7
pika==0.9.14
Pillow==2.9.0
pycrypto==2.6.1
pyflakes==0.8.1
Pygments==2.0.2
pymongo==2.7.1
PyPDF2==1.23
python-dateutil==2.4.2
pytz==2014.9
raven==5.1.1
redis==2.10.3
reportlab==3.1.8
requests==2.7.0
responses==0.5.0
shortuuid==0.4.2
six==1.9.0
snowballstemmer==1.2.0
South==1.0.2
Sphinx==1.3.1
sphinx-me==0.3
sphinx-rtd-theme==0.1.9
sqlparse==0.1.14
ujson==1.33
xhtml2pdf==0.0.6
xlwt==0.7.5
uwsgi-sloth==2.1.2
